Experimental new motion system
change in wheel rotation = change in wheel position = speed...
right?

so now I compute the speed based on the wheel rotation and find
the delta in the given time (which is speed). I compare the desired
and actual speeds to find the delta, and make the force on
the wheel proportional to that.

still *highly* hacky, as the speed will sometimes wildly
increase and decrease.

PID control, anyone? :p